AnimationBranchState
