+++ core/frontend/src/render/GraphicBranch.ts
@@ -8,10 +8,9 @@
 
 import { disposeArray, Id64String } from "@itwin/core-bentley";
 import {
-  FeatureAppearanceProvider, HiddenLine, RealityModelDisplaySettings, RenderSchedule, ViewFlagOverrides, ViewFlags,
+  FeatureAppearanceProvider, HiddenLine, RealityModelDisplaySettings, ViewFlagOverrides, ViewFlags,
 } from "@itwin/core-common";
 import { IModelConnection } from "../IModelConnection";
-import { IModelApp } from "../IModelApp";
 import { FeatureSymbology } from "./FeatureSymbology";
 import { RenderClipVolume } from "./RenderClipVolume";
 import { RenderGraphic } from "./RenderGraphic";
@@ -171,61 +170,4 @@ export interface GraphicBranchOptions {
    * No [ClipStyle.insideColor]($common), [ClipStyle.outsideColor]($common), or [ClipStyle.intersectionStyle]($common) will be applied.
    */
   disableClipStyle?: true;
-}
-
-/** Clip/Transform for a branch that are varied over time.
- * @internal
- */
-export interface AnimationBranchState {
-  readonly clip?: RenderClipVolume;
-  readonly omit?: boolean;
-}
-
-/** @internal */
-export function formatAnimationBranchId(modelId: Id64String, branchId: number): string {
-  if (branchId < 0)
-    return modelId;
-
-  return `${modelId}_Node_${branchId.toString()}`;
-}
-
-function addAnimationBranch(modelId: Id64String, timeline: RenderSchedule.Timeline, branchId: number, branches: Map<string, AnimationBranchState>, time: number): void {
-  const clipVector = timeline.getClipVector(time);
-  const clip = clipVector ? IModelApp.renderSystem.createClipVolume(clipVector) : undefined;
-  if (clip)
-    branches.set(formatAnimationBranchId(modelId, branchId), { clip });
-}
-
-/** Mapping from node/branch IDs to animation branch state
- * @internal
- */
-export interface AnimationBranchStates {
-  /** Maps node Id to branch state. */
-  readonly branchStates: Map<string, AnimationBranchState>;
-  /** Ids of nodes that apply a transform. */
-  readonly transformNodeIds: ReadonlySet<number>;
-}
-
-/** @internal */
-export namespace AnimationBranchStates {
-  export function fromScript(script: RenderSchedule.Script, time: number): AnimationBranchStates | undefined {
-    if (!script.containsModelClipping && !script.requiresBatching)
-      return undefined;
-
-    const branches = new Map<string, AnimationBranchState>();
-    for (const model of script.modelTimelines) {
-      addAnimationBranch(model.modelId, model, -1, branches, time);
-      for (const elem of model.elementTimelines) {
-        if (elem.getVisibility(time) <= 0)
-          branches.set(formatAnimationBranchId(model.modelId, elem.batchId), { omit: true });
-        else
-          addAnimationBranch(model.modelId, elem, elem.batchId, branches, time);
-      }
-    }
-
-    return {
-      branchStates: branches,
-      transformNodeIds: script.transformBatchIds,
-    };
-  }
 }
